# lib: avoid type conversions in encode_json if possible