--- a/lib/json.go
+++ b/lib/json.go
@@ -20,10 +20,7 @@ package lib
 import (
 	"bytes"
 	"encoding/json"
-	"errors"
-	"fmt"
 	"io"
-	"reflect"
 
 	structpb "github.com/golang/protobuf/ptypes/struct"
 	"github.com/google/cel-go/cel"
@@ -32,8 +29,6 @@ import (
 	"github.com/google/cel-go/common/types/ref"
 	"github.com/google/cel-go/interpreter/functions"
 	expr "google.golang.org/genproto/googleapis/api/expr/v1alpha1"
-	"google.golang.org/protobuf/encoding/protojson"
-	"google.golang.org/protobuf/proto"
 )
 
 // JSON returns a cel.EnvOption to configure extended functions for JSON
@@ -193,59 +188,15 @@ func (l jsonLib) ProgramOptions() []cel.ProgramOption {
 }
 
 func encodeJSON(val ref.Val) ref.Val {
-	var (
-		v   interface{}
-		err error
-	)
-	typ, ok := encodableTypes[val.Type()]
-	if ok {
-		v, err = val.ConvertToNative(typ)
-		if err != nil {
-			// This should never happen.
-			panic(fmt.Sprintf("json encode mapping out of sync: %v", err))
-		}
-	} else {
-		for _, typ := range protobufTypes {
-			v, err = val.ConvertToNative(typ)
-			if err != nil {
-				v = nil
-			} else {
-				break
-			}
-		}
+	v, err := val.ConvertToNative(structpbValueType)
+	if err != nil {
+		return types.NewErr("failed proto conversion: %v", err)
 	}
-	if v == nil {
-		return types.NewErr("failed to get native value for JSON")
-	}
-	b, err := json.Marshal(v)
-	errType := &json.UnsupportedTypeError{}
-	switch {
-	case err == nil:
-		return types.String(b)
-	case errors.As(err, &errType):
-		// We could not marshal over ref.Val most likely, so convert
-		// to protobuf as an intermediate and retry.
-		v, err := val.ConvertToNative(reflect.TypeOf(&structpb.Value{}))
-		if err != nil {
-			return types.NewErr("failed proto conversion: %v", err)
-		}
-		b, err := protojson.MarshalOptions{}.Marshal(v.(proto.Message))
-		if err != nil {
-			return types.NewErr("failed native conversion: %v", err)
-		}
-		var res interface{}
-		err = json.Unmarshal(b, &res)
-		if err != nil {
-			return types.NewErr("failed json conversion: %v", err)
-		}
-		b, err = json.Marshal(res)
-		if err != nil {
-			return types.NewErr("failed to marshal value to JSON: %v", err)
-		}
-		return types.String(b)
-	default:
+	b, err := json.Marshal(v.(*structpb.Value).AsInterface())
+	if err != nil {
 		return types.NewErr("failed to marshal value to JSON: %v", err)
 	}
+	return types.String(b)
 }
 
 func (l jsonLib) decodeJSON(val ref.Val) ref.Val {