This takes learning from #36 to simplify and improve the `encode_json` extension.

Depends #36 (want benchmark before merging).

Please take a look.