@efd6 2023-06-26
On the face of it, with benchmarking, this looks like a pass.
```
$ benchstat old.bench new.bench
goos: darwin
goarch: amd64
pkg: github.com/elastic/mito
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
                                  │  old.bench   │              new.bench               │
                                  │    sec/op    │   sec/op     vs base                 │
Mito/hello_world_static-16           613.3n ± 1%   594.5n ± 1%    -3.07% (p=0.000 n=20)
Mito/hello_world_object_static-16    2.647µ ± 2%   2.630µ ± 2%         ~ (p=0.304 n=20)
Mito/nested_static-16                10.33µ ± 2%   10.23µ ± 3%         ~ (p=0.565 n=20)
Mito/encode_json_static-16           22.10µ ± 2%   10.36µ ± 1%   -53.11% (p=0.000 n=20)
Mito/nested_collate_static-16        3.802µ ± 2%   3.830µ ± 2%         ~ (p=0.703 n=20)
Mito/hello_world_state-16            687.7n ± 1%   686.8n ± 1%         ~ (p=0.644 n=20)
Mito/hello_world_object_state-16     2.730µ ± 1%   2.691µ ± 1%    -1.43% (p=0.047 n=20)
Mito/nested_state-16                 9.002µ ± 2%   9.011µ ± 2%         ~ (p=0.698 n=20)
Mito/encode_json_state-16            3.278µ ± 1%   8.871µ ± 1%  +170.66% (p=0.000 n=20)
Mito/nested_collate_list_state-16    12.42µ ± 2%   12.35µ ± 3%         ~ (p=0.963 n=20)
Mito/nested_collate_map_state-16     12.70µ ± 2%   12.60µ ± 3%         ~ (p=0.718 n=20)
Mito/null_net-16                     72.62µ ± 2%   72.48µ ± 1%         ~ (p=0.698 n=20)
Mito/hello_world_net-16              78.97µ ± 2%   78.81µ ± 2%         ~ (p=0.820 n=20)
Mito/hello_world_object_net-16       90.13µ ± 4%   88.05µ ± 2%         ~ (p=0.211 n=20)
Mito/nested_net-16                  100.05µ ± 2%   98.32µ ± 1%    -1.73% (p=0.040 n=20)
Mito/encode_json_null_net-16         78.71µ ± 1%   79.21µ ± 1%         ~ (p=0.547 n=20)
Mito/encode_json_net-16              88.21µ ± 1%   97.08µ ± 2%   +10.06% (p=0.000 n=20)
Mito/nested_collate_list_net-16      102.0µ ± 4%   102.3µ ± 3%         ~ (p=0.904 n=20)
Mito/nested_collate_map_net-16       103.7µ ± 4%   102.0µ ± 1%         ~ (p=0.091 n=20)
geomean                              15.60µ        15.78µ         +1.14%
```
I will look into why there is a cost in the encode_json_state bench.