@efd6 2023-06-26
This appears to come down to inefficiency in the protobuf `.AsMap` implementation.

## memprof old:

```
(pprof) top
Showing nodes accounting for 772.53MB, 95.53% of 808.72MB total
Dropped 87 nodes (cum <= 4.04MB)
Showing top 10 nodes out of 34
      flat  flat%   sum%        cum   cum%
  297.52MB 36.79% 36.79%   399.52MB 49.40%  encoding/json.mapEncoder.encode
     102MB 12.61% 49.40%      102MB 12.61%  reflect.copyVal
   67.50MB  8.35% 57.75%   796.54MB 98.49%  github.com/elastic/mito.run
   65.50MB  8.10% 65.85%   502.52MB 62.14%  github.com/elastic/mito/lib.encodeJSON
   53.50MB  6.62% 72.46%    53.50MB  6.62%  google.golang.org/protobuf/types/known/structpb.NewStringValue (inline)
   44.50MB  5.50% 77.97%       56MB  6.92%  github.com/google/cel-go/common/types.NewStringInterfaceMap (inline)
      43MB  5.32% 83.28%       43MB  5.32%  bytes.(*Buffer).grow
      37MB  4.58% 87.86%   437.02MB 54.04%  encoding/json.Marshal
      32MB  3.96% 91.82%       32MB  3.96%  strings.(*Builder).Write
      30MB  3.71% 95.53%   105.01MB 12.98%  encoding/json.(*Encoder).Encode
```

![old](https://github.com/elastic/mito/assets/90160302/81fff373-2c50-4cae-af79-feeca6863ff6)


## memprof new:

```
(pprof) top
Showing nodes accounting for 1360.19MB, 86.17% of 1578.57MB total
Dropped 77 nodes (cum <= 7.89MB)
Showing top 10 nodes out of 40
      flat  flat%   sum%        cum   cum%
  452.61MB 28.67% 28.67%   461.61MB 29.24%  google.golang.org/protobuf/types/known/structpb.(*Struct).AsMap
  268.55MB 17.01% 45.68%   268.55MB 17.01%  reflect.mapassign_faststr
  129.01MB  8.17% 53.86%   805.08MB 51.00%  github.com/google/cel-go/common/types.(*baseMap).ConvertToNative
  115.51MB  7.32% 61.17%   155.01MB  9.82%  encoding/json.mapEncoder.encode
  100.01MB  6.34% 67.51%   100.01MB  6.34%  google.golang.org/protobuf/types/known/structpb.NewStructValue (inline)
   80.50MB  5.10% 72.61%    80.50MB  5.10%  github.com/google/cel-go/common/types.(*stringIfaceMapAccessor).Iterator
      75MB  4.75% 77.36%   108.01MB  6.84%  github.com/google/cel-go/common/types.NewStringInterfaceMap (inline)
   57.50MB  3.64% 81.00%    57.50MB  3.64%  reflect.makemap
      42MB  2.66% 83.66%       42MB  2.66%  google.golang.org/protobuf/types/known/structpb.NewStringValue (inline)
   39.50MB  2.50% 86.17%    39.50MB  2.50%  reflect.copyVal
```

![new](https://github.com/elastic/mito/assets/90160302/72d9fabf-242a-4e6e-af8f-e7923631a4bd)

Similar patterns are seen in the equivalent static benchmark, so it's not clear why that one shows the opposite performance change. Given that static object conversion to JSON is going to be significantly less common, and can be done by the author literally if needed, I think this should not be merged unless the protobuf performance here improves.